Add Asteroid class
Here in the asteroid branch, I've added the Asteroid class file and updated the A3_GROUP14.pde. In the AsteroidGame class the Asteroid object is declared along with an ArrayList to store the Asteroid objects. The nessercery global variables are also declared and initialised in the constructor method. The methods addAsteroid() and updateAsteroid() control the creation and movement of asteroids. Finally the method calls to these methods in the draw loop displays the asteroids to the screen.